+++ spec/libraries/cache_root_spec.rb
@@ -51,19 +51,5 @@ RSpec.describe Apache2::Cookbook::Helpers do
         expect(subject.default_cache_root).to eq '/var/cache/apache2/proxy'
       end
     end
-
-    context 'with arch' do
-      let(:platform_family) { 'arch' }
-      it 'returns the correct path' do
-        expect(subject.default_cache_root).to eq '/var/cache/httpd/mod_cache_disk'
-      end
-    end
-
-    context 'with freebsd' do
-      let(:platform_family) { 'freebsd' }
-      it 'returns the correct path' do
-        expect(subject.default_cache_root).to eq '/var/cache/apache24/proxy'
-      end
-    end
   end
 end