+++ spec/libraries/mpm_support_spec.rb
@@ -1,4 +1,6 @@
 
+require 'spec_helper'
+
 describe '#mpm_support' do
   recipe do
     log default_mpm
@@ -6,44 +8,28 @@ describe '#mpm_support' do
 
   context 'redhat' do
     platform 'redhat'
-    it { is_expected.to write_log('prefork') }
+    it { is_expected.to write_log('event') }
   end
 
   context 'fedora' do
-    platform 'redhat'
-    it { is_expected.to write_log('prefork') }
+    platform 'fedora'
+    it { is_expected.to write_log('event') }
   end
 
   context 'amazon' do
     platform 'amazon'
-    it { is_expected.to write_log('prefork') }
+    it { is_expected.to write_log('event') }
   end
 
   context 'suse' do
     platform 'suse'
-    it { is_expected.to write_log('prefork') }
+    it { is_expected.to write_log('event') }
   end
 
   context 'debian' do
     platform 'debian'
     it { is_expected.to write_log('event') }
   end
-
-  context 'arch' do
-    platform 'arch'
-    it { is_expected.to write_log('prefork') }
-  end
-
-  context 'freebsd' do
-    platform 'freebsd'
-    it { is_expected.to write_log('prefork') }
-  end
-
-  context 'linuxmint' do
-    platform 'linuxmint'
-    it { is_expected.to write_log('event') }
-  end
-
   context 'ubuntu' do
     platform 'ubuntu'
     it { is_expected.to write_log('event') }