+++ spec/resources/install_spec.rb
@@ -29,15 +29,15 @@ describe 'apache2_install' do
         docroot_dir: '/var/www/html',
         error_log: 'error.log',
         keep_alive: 'On',
-        max_keep_alive_requests: 100,
-        keep_alive_timeout: 5,
+        max_keep_alive_requests: 1000,
+        keep_alive_timeout: 2,
         lock_dir: '/var/lock/apache2',
         log_dir: '/var/log/apache2',
         log_level: 'warn',
         pid_file: '/var/run/apache2/apache2.pid',
         run_dir: '/var/run/apache2',
         server_name: 'localhost',
-        timeout: '300'
+        timeout: '60'
       )
     end
   end
@@ -66,15 +66,15 @@ describe 'apache2_install' do
         docroot_dir: '/var/www/html',
         error_log: 'error.log',
         keep_alive: 'On',
-        max_keep_alive_requests: 100,
-        keep_alive_timeout: 5,
+        max_keep_alive_requests: 1000,
+        keep_alive_timeout: 2,
         lock_dir: '/var/lock/apache2',
         log_dir: '/var/log/apache2',
         log_level: 'warn',
         pid_file: '/var/run/apache2/apache2.pid',
         run_dir: '/var/run/apache2',
         server_name: 'localhost',
-        timeout: '300'
+        timeout: '60'
       )
     end