+++ test/integration/ports/controls/default.rb
@@ -7,6 +7,7 @@ control 'responds on 8080' do
   desc 'Apache should respond on on port 8080'
 
   describe http('http://127.0.0.1:8080') do
-    its('body') { should cmp /Apache/ }
+    its('status') { should be >= 200 }
+    its('status') { should be < 500 }
   end
 end